New version of JSON.lua replaced old version from 1.5 years ago
The new version is verified working, at least on my system running OS
X, (10.11.5) and Lightroom CC (2015.6). It looks like the majority of
changed/new lines are additional documentation, but also significant
bug-fixes/improvements (we can hope) and extra functionality… that just
on a brief look over the code. Anyway, it works with this plugin and
may resolve the reported errors that references the JSON.lua file. (If
not, at least that will tell us something more).