Change variable name autoCheckForExistingKeywords for clarity
autoCheckForExistingKeywords is an unclear action. I actually first
thought it was to check for (i.e. look for/verify) keywords already
associated with an image. What it actually means is to set the checkbox
(i.e. select) the checkboxes for those keywords. The setting has been
renamed to autoSelectExistingKeywords and the variable ‘checkKeyword’
renamed to ‘selectKeyword’.