Refactoring in progress; somewhat broken right now
Currently refactoring around logic that iterates the whole keyword term
tree (should work fine for a flat list, but gathers keywords in
hierarchical list into a table, too). It skips over named branches of
the tree (from settings) that might be useless for Clarifai tagging
(any level term put in the ignore list will be skipped over, along with
child/descendant terms) and it makes use of some new helper functions
necessary for the process. More work will be done to get this perfect,
but it should be good when done and support display of multiple
keywords by the same name (with tooltips to show the difference in
keyword parent/ancestry path).