Working near perfectly now.
The InfoProvider could still take some improvements since layout is not
perfect, but at least fields are not being truncated (hanging off the
edge and parts not visible, as was the case before). Now the keywording
UI is near perfect, with auto-selection of all keywords above the
probability threshold which already exist in the library, and we skip
over parts of the tree that have terms we are not interested in, which
helps performance and to avoid adding unwanted/inappropriate terms.
Seems very performant and near perfect UX now. What would still be nice
would be a way to pop up a larger view of an image while deciding which
keywords to apply and/or a field to add terms that were not identified
by Clarifai, but which come to mind during the process, etc.